requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
python-dotenv==1.1.1
pandas==2.3.0
folium==0.20.0
matplotlib==3.10.3
